+++ packages/workflows/src/definition/product/index.ts
@@ -1,1 +1,3 @@
 export * from "./create-products"
+export * as UpdateProductVariants from "./update-product-variants"
+export * from "./update-products"